adjusted to single-file format
